♻️ Move CompassHierarchyUtility to compass.util
